refactor: store tree entries separately and add metadata field to batch
