+++ x/batching/types/keys.go
@@ -15,6 +15,7 @@ var (
 	CurrentBatchNumberKey = collections.NewPrefix(1)
 	BatchesKeyPrefix      = collections.NewPrefix(2)
 	BatchNumberKeyPrefix  = collections.NewPrefix(3)
-	VotesKeyPrefix        = collections.NewPrefix(4)
-	ParamsKey             = collections.NewPrefix(5)
+	TreeEntriesKeyPrefix  = collections.NewPrefix(4)
+	VotesKeyPrefix        = collections.NewPrefix(5)
+	ParamsKey             = collections.NewPrefix(6)
 )
